requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
Faker==37.5.3
playwright==1.54.0
pytest==8.4.1
pytest-base-url==2.1.0
python-dotenv==1.1.1
python-slugify==8.0.4
requests==2.32.4